requirements.txt
anthropic>=0.28.0
fhir.resources>=7.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0
flask>=2.3.0
flask-cors>=4.0.0
python-json-logger>=2.0.0
